readytodeploy_multimedia

+++ controller/groupController.js
@@ -1,9 +1,10 @@
 const Group=require('../model/group')
 const userGroup=require('../model/usergroup')
 const User=require('../model/user')
-// const Chat=require('../model/chat')
+const Chat=require('../model/chat')
 const sequelize=require('../util/database')
-const {Op}=require('sequelize')
+// const Sequelize=require('sequelize')
+
 
 exports.createGroup=async (req,res,next)=>{
     try{
@@ -16,6 +17,7 @@ exports.createGroup=async (req,res,next)=>{
         userId:req.user.id,
         groupId:result.dataValues.id,
         username:req.user.name,
+        groupname:name
        })
        res.status(200).json(result.dataValues)
     }
@@ -26,11 +28,12 @@ exports.createGroup=async (req,res,next)=>{
 
 exports.getGroup=async (req,res,next)=>{
     try{
-        const groups=await Group.findAll({
+        const groups=await userGroup.findAll({
             where: {
-                userid:req.user.id
+                userId:req.user.id
             }
         })
+        
         res.status(200).json(groups)
     }
     catch(err){
@@ -40,20 +43,28 @@ exports.getGroup=async (req,res,next)=>{
 
 exports.addusertogroup=async (req,res,next)=>{
     try{
-       const {ipt,gid}=req.body
+       const {ipt,gid,gname}=req.body
        const user=await User.findAll({attributes:['id','name'],where: {email:ipt}})
-       const uid=user[0].dataValues.id
-       const found=await userGroup.findAll({where:{userId:uid,groupId:gid}})
-       if(found.length>0)
-       res.status(400).json('user already exists in the group!!')
-       else{
-           userGroup.create({
-            userId:uid,
-            groupId:gid,
-            username:user[0].dataValues.name
-           })
-           res.status(200).json({success:true})
-        }
+       if(user.length==0){
+        res.status(201).json('User doesnot exist!')
+       }
+       else if(user.length==1){
+           const uid=user[0].dataValues.id
+           console.log(uid)
+           const found=await userGroup.findAll({where:{userId:uid,groupId:gid}})
+           if(found.length>0)
+           res.status(201).json('user already exists in the group!!')
+           else{
+               userGroup.create({
+                userId:uid,
+                groupId:gid,
+                username:user[0].dataValues.name,
+                groupname:gname
+               })
+               res.status(200).json({success:true})
+            }
+
+       }
        
     }
     catch(err){
@@ -62,15 +73,27 @@ exports.addusertogroup=async (req,res,next)=>{
 }
 
 exports.deleteGroup=async (req,res,next)=>{
+    const t=await sequelize.transaction()  
     try{
+      const uid=req.user.id
       const gid=req.params.id
-      await Group.destroy({where: {id:gid}})
-      await userGroup.destroy({where: {groupId:gid}})
-    //   await Chat.destroy({where: {groupId:gid}})
-      res.status(200).json('deleted')
+      const found=await Group.findOne({where: {id:gid,userid:uid}})
+      if(found){
+          await Group.destroy({where: {id:gid}},{ transaction: t })
+          await userGroup.destroy({where: {groupId:gid}},{ transaction: t })
+          await Chat.destroy({where: {groupId:gid}},{ transaction: t })
+          await t.commit()
+          res.status(200).json('deleted')
+        }
+        else{
+            await t.commit()
+            res.status(400).json('You are NOT the admin!')
+        }
     }
     catch(err){
+        await t.rollback()
         console.log(err)
+        res.status(500).json('Something went wrong!')
     }
 }
 
@@ -83,7 +106,6 @@ exports.getGrpUsers=async (req,res,next)=>{
         for(let i=0;i<grpusers.length;i++){
             users.push(grpusers[i].dataValues)
         }
-        console.log(users)
         res.status(200).json(users)
     }
     catch(err){
@@ -93,12 +115,24 @@ exports.getGrpUsers=async (req,res,next)=>{
 
 exports.removeUserFromGroup=async (req,res,next)=>{
     try{
+      console.log('reached!!')
+        const ownuid=req.user.id
+        console.log(ownuid)
       const {groupid,userid}=req.query
-      await userGroup.destroy({where: {groupId:groupid,userId:userid}})
-      res.status(200).json({success:true})
+      console.log(userid)
+      const found=await Group.findOne({where: {id:groupid,userid:userid}})
+      console.log(found)
+      if(found){
+            await userGroup.destroy({where: {groupId:groupid,userId:userid}})
+            res.status(200).json('User removed succefully!')
+      }
+      else{
+            res.status(400).json('You are NOT the admin!')
+        }
     }
     catch(err){
       console.log(err)
+      res.status(500).json('Something went wrong!')
     }
 }
